+++ extractor/src/test/suite/scala/com/digitalasset/extractor/RecordsAndVariantsSpec.scala
@@ -3,17 +3,15 @@
 
 package com.digitalasset.extractor
 
+import java.io.File
+
 import com.digitalasset.daml.bazeltools.BazelRunfiles._
 import com.digitalasset.extractor.services.{CustomMatchers, ExtractorFixtureAroundAll}
 import com.digitalasset.ledger.api.testing.utils.SuiteResourceManagementAroundAll
-import com.digitalasset.platform.sandbox.persistence.PostgresAroundAll
-
-import org.scalatest._
+import com.digitalasset.testing.postgresql.PostgresAroundAll
 import io.circe.parser._
-import java.io.File
-
-import scalaz._
-import Scalaz._
+import org.scalatest._
+import scalaz.Scalaz._
 
 @SuppressWarnings(Array("org.wartremover.warts.Any"))
 class RecordsAndVariantsSpec
