+++ extractor/src/test/suite/scala/com/digitalasset/extractor/TemplateSubscriptionSpec.scala
@@ -10,7 +10,7 @@ import com.digitalasset.daml.lf.data.Ref.Party
 import com.digitalasset.extractor.config.{ExtractorConfig, TemplateConfig}
 import com.digitalasset.extractor.services.ExtractorFixtureAroundAll
 import com.digitalasset.ledger.api.testing.utils.SuiteResourceManagementAroundAll
-import com.digitalasset.platform.sandbox.persistence.PostgresAroundAll
+import com.digitalasset.testing.postgresql.PostgresAroundAll
 import org.scalatest.{FlatSpec, Inside, Matchers, Suite}
 import scalaz.OneAnd
