+++ extractor/src/test/suite/scala/com/digitalasset/extractor/TransactionMultiTableSpec.scala
@@ -11,10 +11,9 @@ import cats.implicits._
 import com.digitalasset.daml.bazeltools.BazelRunfiles._
 import com.digitalasset.extractor.services.{CustomMatchers, ExtractorFixtureAroundAll}
 import com.digitalasset.ledger.api.testing.utils.SuiteResourceManagementAroundAll
-import com.digitalasset.platform.sandbox.persistence.PostgresAroundAll
-import io.circe.syntax._
+import com.digitalasset.testing.postgresql.PostgresAroundAll
 import doobie.implicits._
-
+import io.circe.syntax._
 import org.scalatest._
 
 import scala.concurrent.duration._
