+++ ledger-service/http-json/BUILD.bazel
@@ -136,8 +136,8 @@ da_scala_test(
         "//ledger/ledger-api-common",
         "//ledger/participant-state",
         "//ledger/sandbox",
-        "//ledger/sandbox:sandbox-scala-tests-lib",
         "//libs-scala/auth-utils",
+        "//libs-scala/postgresql-testing",
         "@maven//:ch_qos_logback_logback_classic",
         "@maven//:com_chuusai_shapeless_2_12",
         "@maven//:com_lihaoyi_sourcecode_2_12",
