+++ ledger/sandbox-perf/BUILD.bazel
@@ -33,6 +33,7 @@ da_scala_library(
         "//ledger/sandbox:sandbox-scala-tests-lib",
         "//ledger/test-common",
         "//libs-scala/direct-execution-context",
+        "//libs-scala/postgresql-testing",
         "@maven//:ch_qos_logback_logback_classic",
         "@maven//:com_google_protobuf_protobuf_java",
         "@maven//:com_typesafe_akka_akka_actor_2_12",
