+++ ledger/sandbox/BUILD.bazel
@@ -261,6 +261,7 @@ da_scala_library(
         "//libs-scala/contextualized-logging",
         "//libs-scala/direct-execution-context",
         "//libs-scala/grpc-utils",
+        "//libs-scala/postgresql-testing",
         "//libs-scala/resources",
         "//libs-scala/timer-utils",
         "@maven//:ch_qos_logback_logback_classic",
@@ -270,7 +271,6 @@ da_scala_library(
         "@maven//:com_typesafe_akka_akka_stream_2_12",
         "@maven//:com_typesafe_config",
         "@maven//:com_typesafe_scala_logging_scala_logging_2_12",
-        "@maven//:commons_io_commons_io",
         "@maven//:io_dropwizard_metrics_metrics_core",
         "@maven//:io_grpc_grpc_netty",
         "@maven//:io_netty_netty_common",
@@ -318,6 +318,7 @@ da_scala_test_suite(
         "//libs-scala/contextualized-logging",
         "//libs-scala/direct-execution-context",
         "//libs-scala/grpc-utils",
+        "//libs-scala/postgresql-testing",
         "//libs-scala/resources",
         "//libs-scala/timer-utils",
         "@maven//:ch_qos_logback_logback_classic",
